TT move ordering
Results of Cairn_new vs Cairn_old (10+0.1, 1t - NULL, 16MB, 8moves_v3.pgn):
Elo: 175.40 +/- 45.61, nElo: 234.88 +/- 51.33
LOS: 100.00 %, DrawRatio: 23.86 %, PairsRatio: 12.40
Games: 176, Wins: 106, Losses: 24, Draws: 46, Points: 129.0 (73.30 %)
Ptnml(0-2): [3, 2, 21, 34, 28], WL/DD Ratio: 3.20
LLR: 2.22 (101.2%) (-2.20, 2.20) [0.00, 10.00]

+++ src/preComputedData.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 #define TABLE_SIZE 64
-extern unsigned long long knight_moves[TABLE_SIZE];
-extern unsigned long long king_moves[TABLE_SIZE];
+extern uint64_t knight_moves[TABLE_SIZE];
+extern uint64_t king_moves[TABLE_SIZE];
 extern const uint64_t a_file;
 extern const uint64_t first_rank;
 extern uint64_t rays[TABLE_SIZE][8];
